Feat SCRUM-28: route handler 검증을 위한 userInfo API 로직 개발
- 단순 api 구현, 추후 react-query로 리팩토링 예정
