+++ app/store/types/user.types.ts
@@ -1,0 +1,10 @@
+export interface User {
+  teamId: number;
+  id: number;
+  email: string;
+  name: string;
+  companyName: string;
+  image: string;
+  createdAt: string;
+  updatedAt: string;
+}
